requirements.txt
glob
re
cv2
numpy
sklearn.metrics
cv2
import shutil
